Remove Type module and Show instance and use deriving.
